# docs/SetupDocs/nextxr_openapi.yaml
openapi: 3.0.3
info:
  title: NextXR Digital Twin Platform API
  description: |
    Ontology-driven industrial digital-twin API. Covers twin lifecycle, the
    semantic graph and its governance, the machine-twin physics runtime,
    telemetry ingestion and history, field-protocol connectors, solar PV
    analytics, the embedded AI copilot, agentic twin authoring, 3-D
    reconstruction, the live event bus and platform health.

    136 endpoints across 12 routers plus a mounted 3-D sub-application.

    ## Authentication

    Two credential types, deliberately not interchangeable.

    - **`X-API-Key`** — a human/service credential carrying a **role**
      (`admin`, `write`, `read`) and a **tenant scope**. Configured as a JSON
      array in `NXR_API_KEYS`; the built-in demo keys load **only** when that
      variable is unset (i.e. never in a correctly-configured deployment).
      Set `NXR_REQUIRE_AUTH=1` to reject keyless calls even before keys load.
    - **`X-Device-Token`** — a telemetry device credential granting exactly one
      verb (append telemetry) in exactly one tenant, optionally narrowed to an
      asset prefix. Returned once at registration; only the hash is stored.
      It is never accepted on a read, control or copilot route, and an API key
      is never accepted on the ingest routes in its place.

    > While `NXR_API_KEYS` is unset the API is **OPEN** — every `/api` call,
    > including the LLM-billing `/copilot` endpoints, is served without a key.

...

...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/schema/class/{name}:
    get:
      tags: [schema]
      summary: Class definition and metadata
      operationId: schemaClass
      parameters: [ { $ref: '#/components/parameters/ClassName' } ]
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/schema/class/{name}/properties:
    get:
      tags: [schema]
      summary: Observable and object properties of a class
      operationId: schemaClassProperties
      parameters: [ { $ref: '#/components/parameters/ClassName' } ]
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
  /api/v1/schema/class/{name}/behavior:
    get:
      tags: [schema]
      summary: Behaviours bound to a class
      operationId: schemaClassBehavior
      parameters: [ { $ref: '#/components/parameters/ClassName' } ]
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
...
  /api/v1/schema/validate:
    post:
      tags: [schema]
      summary: SHACL-validate a payload without writing it
      operationId: schemaValidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [canonical_type, properties]
              properties:
                canonical_type: { type: string }
                properties: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Conformance report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValidationReport' }
...
  /api/v1/twins:
    get:
      tags: [twins]
      summary: List twins in the caller's scope
      operationId: listTwins
      responses:
        '200':
          description: Twin registry records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Twin' }
    post:
      tags: [twins]
      summary: Create and seed a twin
      description: |
        Entities are seeded through the Graph Writer, so validation, change log
        and bus events all fire. Only a prefix-scoped or admin key may create a
        twin — the new twin's id has to land inside the creator's own scope to
        remain readable.
      operationId: createTwin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id, name, domain]
              properties:
                tenant_id: { type: string }
                name: { type: string }
                domain:
                  type: string
                  description: Seeding template or machine-twin domain key
                  example: solar-pv-array
                description: { type: string }
      responses:
        '201':
          description: Created and seeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Twin' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/twins/{tenant}:
    parameters: [ { $ref: '#/components/parameters/TenantPath' } ]
    get:
      tags: [twins]
      summary: Twin registry record
      operationId: getTwin
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Twin' }
          description: The twin
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [twins]
      summary: Delete a twin and its entities
      description: The change log is retained as an audit record of what existed.
      operationId: deleteTwin
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ack' }
          description: Deleted

  # ==========================================================================
  # twin-runtime
  # ==========================================================================
...
  /api/v1/twins/{tenant}/state:
    get:
      tags: [twin-runtime]
      summary: Live frame — signals, control, running flag, health, faults
      operationId: twinState
      parameters: [ { $ref: '#/components/parameters/TenantPath' } ]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TwinState' }
          description: The authoritative state published by the lease owner
  /api/v1/twins/{tenant}/diagnostics:
    get:
      tags: [twin-runtime]
      summary: Per-sensor status and per-subsystem health
      operationId: twinDiagnostics
      parameters: [ { $ref: '#/components/parameters/TenantPath' } ]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Diagnostics' }
          description: Evaluated from the pack's checks table and component_health
  /api/v1/twins/{tenant}/predict:
    get:
      tags: [twin-runtime]
      summary: Remaining useful life and degradation forecast
      operationId: twinPredict
      parameters: [ { $ref: '#/components/parameters/TenantPath' } ]
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
  /api/v1/twins/{tenant}/project:
    post:
      tags: [twin-runtime]
      summary: What-if projection (does not mutate the live twin)
      operationId: twinProject
      parameters: [ { $ref: '#/components/parameters/TenantPath' } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                control: { type: number }
                horizon_hours: { type: number }
                fault: { type: string }
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
  /api/v1/twins/{tenant}/network:
    get:
      tags: [twin-runtime]
      summary: Network state for network-shaped domains
      description: Metro, tram, EV charging and defence-base twins. Others return 404.
      operationId: twinNetwork
      parameters: [ { $ref: '#/components/parameters/TenantPath' } ]
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/twins/{tenant}/running:
    post:
      tags: [twin-runtime]
      summary: Start or stop the simulation
      description: A follower task forwards the command to the lease owner rather than acting locally.
      operationId: twinRunning
      parameters: [ { $ref: '#/components/parameters/TenantPath' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [running]
              properties:
                running: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
  /api/v1/twins/{tenant}/simulate:
    post:
      tags: [twin-runtime]
      summary: Inject a fault or set the control input
      description: |
        The fault persists in the stateful integrator until cleared — wear
        accumulates and heat soaks, so this is a simulation, not a value override.
      operationId: twinSimulate
      parameters: [ { $ref: '#/components/parameters/TenantPath' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fault:
                  type: string
                  description: A key from the pack's FAULTS table
                control:
                  type: number
                  description: The pack's declared control input (throttle, curtailment, ...)
                clear: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }

  # ==========================================================================
  # ingest
  # ==========================================================================
...
  /api/v1/agents/twin/{session_id}:
    get:
      tags: [agents]
      summary: Twin-builder session state
      operationId: twinSessionState
      parameters: [ { $ref: '#/components/parameters/SessionId' } ]
      responses:
        '200': { $ref: '#/components/responses/AgentSession' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Carries a role (`admin` / `write` / `read`) and a tenant scope. Configured
        via `NXR_API_KEYS`.
    DeviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
      description: |
        A telemetry device credential: one verb (append), one tenant, optionally
        one asset prefix. Returned once at registration; only the hash is stored.
  parameters:
...
    TenantPath:
      name: tenant
      in: path
      required: true
      schema: { type: string }
      description: Tenant id; checked against the caller's key scope after routing
...
    ClassName:
      name: name
      in: path
      required: true
      schema: { type: string }
      description: Ontology class local name (e.g. AirHandler)
...
    SessionId:
      name: session_id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    JsonObject:
      description: Success
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
...
    AgentSession:
      description: Agent graph session state, checkpointed by thread_id
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
...
    Forbidden:
      description: Tenant outside the caller's key scope, or role insufficient
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Ack:
      type: object
      properties:
        ok: { type: boolean }
        detail: { type: string }
    Error:
      type: object
      properties:
        detail: { type: string }
...
    Finding:
      type: object
      properties:
        id: { type: string }
        behaviourId: { type: string, example: solar.soiling_symmetric_drop }
        tier:
          type: string
          enum: [A, B, C]
          description: A = physics residual, B = statistical baseline, C = deterministic rule
        severity: { type: string, enum: [info, warning, critical] }
        entityId: { type: string }
        evidence: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
...
    Twin:
      type: object
      properties:
        tenant_id: { type: string }
        name: { type: string }
        domain: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
        seed_asset_id: { type: string }
...
    TwinState:
      type: object
      properties:
        tenant: { type: string }
        domain: { type: string }
        running: { type: boolean }
        control: { type: number }
        health_index: { type: number }
        faults: { type: array, items: { type: string } }
        signals:
          type: object
          additionalProperties:
            type: object
            properties:
              value: { type: number }
              unit: { type: string }
        ts: { type: string, format: date-time }
    Diagnostics:
      type: object
      properties:
        sensors:
          type: array
          items:
            type: object
            properties:
              signal: { type: string }
              value: { type: number }
              limit: { type: number }
              direction: { type: string, enum: [above, below] }
              status: { type: string, enum: [ok, warning, alarm] }
        subsystems:
          type: array
          items:
            type: object
            properties:
              key: { type: string }
              label: { type: string }
              health: { type: number }
        findings:
          type: array
          items: { $ref: '#/components/schemas/Finding' }
...
    ValidationReport:
      type: object
      properties:
        conforms: { type: boolean }
        violations:
          type: array
          items:
            type: object
            properties:
              shape: { type: string }
              focus_node: { type: string }
              path: { type: string }
              message: { type: string }
              severity: { type: string }